requirements.txt
Flask
psycopg2-binary
sqlalchemy>=1.4
alembic
ics
apscheduler
python-dotenv
